Allow word-characters and underscore process type names
A Procfile entry doesn't need to end in _worker, _handler, etc… to be
valid. For example, `release` is a special entry in a Heroku Procfile,
but it shows up as invalid. Similarly names like `sidekiq` or
`stream_consumer` are valid.

This loosens the regex to require at least one word character, and then
optionally any number of underscores or word characters, ending in a
colon.